imagem vindo do banco ok

+++ pi4_fellas.coffe-main/SpringWeb.v1/src/main/java/com/web/pi3s/SpringWeb/controllers/ProdutoController.java
@@ -131,7 +131,7 @@ public class ProdutoController {
     }
 
     @GetMapping("/product/{id}")
-    public ResponseEntity<byte[]> exibirImagem(@PathVariable Long id) {
+    public ResponseEntity<byte[]> exibirImagem(@PathVariable int id) {
         Optional<Imagenmodels> imagemOptional = imagenrespo.findById(id);
 
         if (imagemOptional.isPresent()) {
